Add search.py; Youtube and OMDB API intigration
This commit adds the search module with 2 functions. youtube_search
intigrates the Youtube API service to search for a movie trailer on
youtube.com. omdb_search placess HTTP GET request to omdbapi.com
using the Requests library to search for movie metadata.

These functions, given an api key, return another function that will be
used in the Movie class as a static method to help search for movie info.